Move DB entities to entity package
