+++ src/main/java/com/paddavoet/bittradr/repository/TradeVelocityRepository.java
@@ -1,6 +1,6 @@
 package com.paddavoet.bittradr.repository;
 
-import com.paddavoet.bittradr.market.TradeVelocity;
+import com.paddavoet.bittradr.entity.TradeVelocity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
